Assignment 2/requirements.txt
# Core ML Libraries
torch>=2.2.0
torchvision>=0.17.0
scikit-learn>=1.3.2
numpy>=1.24.3
pandas>=2.0.3

# Experiment Tracking
mlflow>=2.8.1

# API Framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6

# Testing
pytest>=7.4.3
pytest-cov>=4.1.0
httpx>=0.25.1

# DVC for Data Versioning
dvc>=3.30.1

# Monitoring & Logging
prometheus-client>=0.19.0
python-json-logger>=2.0.7

# Data Download
kaggle>=1.5.16

# Utilities
pillow>=10.1.0
matplotlib>=3.8.1
seaborn>=0.13.0
tqdm>=4.66.1
